[Chore] Move keyboard extension to UIViewController+Extension file
