+++ Gwayeon/Globals/Extension/UIViewController+Extension.swift
@@ -1,0 +1,20 @@
+//
+//  UIViewController+Extension.swift
+//  Gwayeon
+//
+//  Created by 이윤영 on 2022/07/28.
+//
+
+import UIKit
+
+extension UIViewController {
+    func hideKeyboardWhenTappedAround() {
+        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        view.addGestureRecognizer(tap)
+    }
+    
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
