+++ Gwayeon/RecommendFarmViewController.swift
@@ -88,16 +88,3 @@ class RecommendFarmViewController: UIViewController {
         }
     }
 }
-
-// TODO: Extension 폴더 분리
-extension UIViewController {
-    func hideKeyboardWhenTappedAround() {
-        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
-        tap.cancelsTouchesInView = false
-        view.addGestureRecognizer(tap)
-    }
-    
-    @objc func dismissKeyboard() {
-        view.endEditing(true)
-    }
-}
